requirements.txt
######################################## Requirements file ########################################

######################################## Python requirements ########################################
numpy
scipy
Pillow
grpcio
grpcio-tools

